// backend/src/models/TituloJogo.ts
export class TituloJogo{
    id: number = 0;
    nome: string = '';
    descricao: string = '';
    data_lancamento:  string = '';
    nota: number = 0;
    generos: string[] = [];
    plataformas: string[] = [];
    produtoras: string[] = [];
    poster: string = 'https://images.igdb.com/igdb/image/upload/t_cover_big_2x/';
    website: string[] = [];
}